fix address_label_token_balance_staked address_label_token_time_first_stake

+++ app/tag-app/src/main/resources/initsql/address_label_gp.sql
@@ -60,7 +60,7 @@ create table address_label_gp_${tableSuffix}
     project varchar(80) NULL,
     asset varchar(80) NULL,
     bus_type varchar(100) NULL,
-    id bigint
+    id int8
 ) distributed by (address);
